feat: Add initial home page structure with Hero, Features, How It Works, and Footer components.
