+++ web/src/components/home/index.ts
@@ -1,0 +1,4 @@
+export { Hero } from './Hero';
+export { Features } from './Features';
+export { HowItWorks } from './HowItWorks';
+export { Footer } from './Footer';
